+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
-*no unreleased changes*
+### Fixed
+* Capistrano: deploy:preinstall should fix up installed gem permissions
+* Support Ruby 3.4
+
+### Added
+* Capistrano: cleanup bundled gems for old ruby versions.
 
 ## 7.3.3 / 2025-07-31
 ### Fixed